refactor: update vue-renderer-markdown to 0.0.54-beta.6 and remove ReferenceNode component
