+++ package.json
@@ -157,7 +157,7 @@
     "vitest": "^3.2.4",
     "vue": "^3.5.21",
     "vue-i18n": "^11.1.11",
-    "vue-renderer-markdown": "0.0.54-beta.4",
+    "vue-renderer-markdown": "0.0.54-beta.6",
     "vue-router": "4",
     "vue-tsc": "^2.2.12",
     "vue-use-monaco": "^0.0.31",
